fix function sharing
