+++ tests/utils.go
@@ -1,0 +1,57 @@
+// The MIT License
+//
+// Copyright (c) 2020 Temporal Technologies Inc.  All rights reserved.
+//
+// Copyright (c) 2020 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package tests
+
+import (
+	historypb "go.temporal.io/api/history/v1"
+	historyspb "go.temporal.io/server/api/history/v1"
+	"go.temporal.io/server/common/persistence/versionhistory"
+)
+
+func EventBatchesToVersionHistory(
+	versionHistory *historyspb.VersionHistory,
+	eventBatches []*historypb.History,
+) (*historyspb.VersionHistory, error) {
+
+	// TODO temporary code to generate version history
+	//  we should generate version as part of modeled based testing
+	if versionHistory == nil {
+		versionHistory = versionhistory.NewVersionHistory(nil, nil)
+	}
+	for _, batch := range eventBatches {
+		for _, event := range batch.Events {
+			err := versionhistory.AddOrUpdateVersionHistoryItem(versionHistory,
+				versionhistory.NewVersionHistoryItem(
+					event.GetEventId(),
+					event.GetVersion(),
+				))
+			if err != nil {
+				return nil, err
+			}
+		}
+	}
+
+	return versionHistory, nil
+}
